pyproject.toml
[tool.poetry]
name = "datasaude-api-ml"
version = "0.1.0"
description = ""
authors = ["caue beloni <cbeloni@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pandas = "==2.2.2"
numpy = "==1.26.4"
matplotlib = "==3.8.4"
scikit-learn = "==1.4.2"
seaborn = "==0.13.2"
openpyxl = "==3.1.2"
joblib = "==1.4.0"
xgboost = "==2.0.3"
imblearn = "==0.0"
statsmodels = "==0.14.2"
sqlalchemy = "^2.0.31"
aiomysql = "^0.2.0"
databases = "^0.9.0"
pymysql = "^1.1.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
